requirements.txt
asgiref==3.5.2; python_version >= "3.8"
backports.zoneinfo==0.2.1; python_version < "3.9" and python_version >= "3.8"
certifi==2022.6.15; python_version >= "3.7" and python_version < "4"
charset-normalizer==2.1.0; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0"
coreapi==2.3.3; python_version >= "3.6"
coreschema==0.0.4; python_version >= "3.6"
django-cors-headers==3.13.0; python_version >= "3.7"
django-extensions==3.2.0; python_version >= "3.6"
django-restframework==0.0.1
django==4.0.6; python_version >= "3.8"
djangorestframework-simplejwt==5.2.0; python_version >= "3.7"
djangorestframework==3.13.1; python_version >= "3.7"
drf-yasg==1.21.3; python_version >= "3.6"
gunicorn==20.1.0; python_version >= "3.5"
idna==3.3; python_version >= "3.7" and python_version < "4"
inflection==0.5.1; python_version >= "3.6"
itypes==1.2.0; python_version >= "3.6"
jinja2==3.1.2; python_version >= "3.7"
markupsafe==2.1.1; python_version >= "3.7"
mysqlclient==2.1.1; python_version >= "3.5"
packaging==21.3; python_version >= "3.6"
pyjwt==2.4.0; python_version >= "3.7"
pyparsing==3.0.9; python_full_version >= "3.6.8" and python_version >= "3.6"
python-dotenv==0.20.0; python_version >= "3.5"
pytz==2022.1; python_version >= "3.7"
requests==2.28.1; python_version >= "3.7" and python_version < "4"
ruamel.yaml.clib==0.2.6; platform_python_implementation == "CPython" and python_version < "3.11" and python_version >= "3.6"
ruamel.yaml==0.17.21; python_version >= "3.6"
sqlparse==0.4.2; python_version >= "3.8"
tzdata==2022.1; sys_platform == "win32" and python_version >= "3.8"
uritemplate==4.1.1; python_version >= "3.6"
urllib3==1.26.10; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7"
